# openapi/spec.yml
openapi: 3.0.0
info:
  title: Product Inventory API
  version: 1.0.0
paths:
  /products:
    get:
      summary: Get all products
      responses:
        '200':
          description: Fetch the full list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Add a new product to the inventory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /orders:
    get:
      summary: Get all orders
      responses:
        '200':
          description: Fetch the full list of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /process_order:
    post:
      summary: Process an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/Order'
      responses:
        '201':
          description: Processed order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /process_restock:
    post:
      summary: Restocks a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/Restock'
      responses:
        '201':
          description: Restocked product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restock'
  /ship_package:
    post:
      summary: Ship a /ship_package
      requestBody:
        required: true
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/Shipment'
      responses:
        '201':
          description: Shipped package
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shipment'
components:
  schemas:
    Product:
      type: object
      properties:
        product_id:
          type: number
        product_name: 
          type: string
        mass_g: 
          type: integer
      required:
        - product_id
        - product_name
        - mass_g
    OrderLineItem:
      type: object
      properties:
        product_id:
          type: number
        quantity:
          type: number
      required:
        - product_id
        - quantity
    Order:
      type: object
      properties:
        order_id:
          type: number
        requested:
          type: array
          items: 
            $ref: '#/components/schemas/OrderLineItem'
      required:
        - order_id
        - requested
    Shipment:
      type: object
      properties:
        order_id:
          type: number
        shipped:
          type: array
          items:
            $ref: '#/components/schemas/OrderLineItem'
    Restock:
      type: object
      properties:
        product_id:
          type: number
        quantity:
          type: number
      required:
        - product_id
        - quantity